avoid multiple LLM instantiation, remove single batch example.
