+++ vllm_batch_inference/batch_inference.py
@@ -1,19 +1,29 @@
 import json
 from pathlib import Path
+from typing import Generator
 
-import vllm
+from vllm import LLM, SamplingParams
 
 
-def load_data(file: Path) -> list[dict[str, str]]:
-    """Load data from a jsonl file. Assuming at least `id` and `input` fields are present.
+def load_data(
+    file: Path, batch_size: int
+) -> Generator[list[dict[str, str]], None, None]:
+    """Load data from a jsonl file as a generator. Assuming at least `id` and `input` fields are present.
 
     Example input format: (docs: https://jsonlines.org/)
-    {"id": "x0001", input": "What is the capital of France?"}
-    {"id": "x0002", input": "What is the capital of Germany?"}
+    {"id": "q0001", "input": "What is the capital of France?"}
+    {"id": "q0002", "input": "What is the capital of Germany?"}
     """
 
     with open(file, mode="r") as f:
-        return [json.loads(line) for line in f]
+        batch = []
+        for line in f:
+            batch.append(json.loads(line))
+            if len(batch) == batch_size:
+                yield batch
+                batch.clear()
+        if batch:
+            yield batch
 
 
 def format_prompt(user_message: str, system_message: str | None = None) -> str:
@@ -24,7 +34,7 @@ def format_prompt(user_message: str, system_message: str | None = None) -> str:
 
     if system_message is None:
         system_message = (
-            "You are a helpful assistant. Provide clear and concise responses."
+            "You are a helpful assistant. Provide clear and short responses."
         )
     return (
         "<|system|>\n"
@@ -35,7 +45,13 @@ def format_prompt(user_message: str, system_message: str | None = None) -> str:
     )
 
 
-def inference(data: list[dict[str, str]], input_field: str = "input") -> list[dict]:
+def inference(
+    llm: LLM,
+    data: list[dict[str, str]],
+    input_field: str = "input",
+    temperature: float = 0.0,
+    max_tokens: int = 2048,
+) -> list[dict]:
     """Perform a batch of inference on Phi-3.5-mini-instruct via vllm offline mode.
 
     docs: https://docs.vllm.ai/en/v0.6.4/getting_started/examples/offline_inference.html
@@ -44,10 +60,7 @@ def inference(data: list[dict[str, str]], input_field: str = "input") -> list[di
     inputs = [item[input_field] for item in data]
     formatted_inputs = [format_prompt(user_message=x) for x in inputs]
 
-    llm = vllm.LLM(model="microsoft/Phi-3.5-mini-instruct")
-
-    # Set sampling parameters
-    sampling_params = vllm.SamplingParams(temperature=0, max_tokens=2048)
+    sampling_params = SamplingParams(temperature=temperature, max_tokens=max_tokens)
     outputs = llm.generate(prompts=formatted_inputs, sampling_params=sampling_params)
     text_outputs = [raw_output.outputs[0].text.strip() for raw_output in outputs]
     return [{**item, "output": output} for item, output in zip(data, text_outputs)]
@@ -57,15 +70,22 @@ def save_data(data: list[dict], file: Path) -> None:
     """Save data to a jsonl file in append mode.
 
     Example output format:
-    {"id": "x0001", "input": "What is the capital of France?", "output": "Paris."}
-    {"id": "x0002", "input": "What is the capital of Germany?", "output": "Berlin."}
+    {"id": "q0001", "input": "What is the capital of France?", "output": "Paris."}
+    {"id": "q0002", "input": "What is the capital of Germany?", "output": "Berlin."}
     """
     with open(file, mode="a") as f:
         for item in data:
             f.write(json.dumps(item) + "\n")
 
 
-# Load data, perform inference, and save data
-data = load_data(Path("inputs.jsonl"))
-outputs = inference(data)
-save_data(outputs, Path("outputs.jsonl"))
+# Main script
+# Perform mini-batch inference on the input data and save the results.
+# Adjust the batch size based on your requirements.
+# Note: Larger batch sizes may require more GPU memory but can be faster.
+
+data = load_data(Path("inputs.jsonl"), batch_size=20)
+llm = LLM(model="microsoft/Phi-3.5-mini-instruct", max_model_len=8192)
+for batch in data:
+    outputs = inference(llm=llm, data=batch)
+    save_data(outputs, Path("outputs.jsonl"))
+    print(f"Processed {len(batch)} examples.")
